finetuning_evaluation/setup_casia_webface/requirements.txt
mxnet-mkl==1.6.0
numpy==1.23.1
tqdm
pillow
opencv-python
argparse
